requirements.txt
mcp[cli]>=1.26.0
anthropic>=0.49.0
httpx>=0.27.0
uvicorn>=0.30.0
a2a-sdk[http-server]>=0.3.22
python-dotenv>=1.0.0
